+++ src/dns-resolver/azext_dnsresolver/aaz/latest/dns_resolver/policy/dns_security_rule/_list.py
@@ -22,9 +22,9 @@ class List(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2023-07-01-preview",
+        "version": "2025-05-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.network/dnsresolverpolicies/{}/dnssecurityrules", "2023-07-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.network/dnsresolverpolicies/{}/dnssecurityrules", "2025-05-01"],
         ]
     }
 
@@ -133,7 +133,7 @@ class List(AAZCommand):
                     "$top", self.ctx.args.top,
                 ),
                 **self.serialize_query_param(
-                    "api-version", "2023-07-01-preview",
+                    "api-version", "2025-05-01",
                     required=True,
                 ),
             }
@@ -222,9 +222,6 @@ class List(AAZCommand):
             action = cls._schema_on_200.value.Element.properties.action
             action.action_type = AAZStrType(
                 serialized_name="actionType",
-            )
-            action.block_response_code = AAZStrType(
-                serialized_name="blockResponseCode",
             )
 
             dns_resolver_domain_lists = cls._schema_on_200.value.Element.properties.dns_resolver_domain_lists