+++ src/dns-resolver/azext_dnsresolver/aaz/latest/dns_resolver/policy/dns_security_rule/_update.py
@@ -17,14 +17,14 @@ from azure.cli.core.aaz import *
 class Update(AAZCommand):
     """Update a DNS security rule for a DNS resolver policy.
 
-    :example: Update a DNS security rule
-        az dns-resolver policy dns-security-rule update --resource-group sampleResourceGroup --policy-name sampleDnsResolverPolicy --rule-name sampleDnsSecurityRule --location westus2 --tags "{key2:value2}" --priority 100 --action "{action-type:Block,block-response-code:SERVFAIL}" --domain-lists "[{id:/subscriptions/abdd4249-9f34-4cc6-8e42-c2e32110603e/resourceGroups/sampleResourceGroup/providers/Microsoft.Network/dnsResolverDomainLists/sampleDnsResolverDomainList}]" --rule-state Disabled
+    :example: Update a dns-security rule.
+        az dns-resolver policy dns-security-rule update --resource-group sampleResourceGroup --policy-name sampleDnsResolverPolicy --dns-security-rule-name sampleDnsSecurityRule --location westus2 --tags "{key1:value1}" --priority 100 --action "{action-type:Block}" --domain-lists "[{id:/subscriptions/abdd4249-9f34-4cc6-8e42-c2e32110603e/resourceGroups/sampleResourceGroup/providers/Microsoft.Network/dnsResolverDomainLists/sampleDnsResolverDomainList}]" --rule-state Enabled
     """
 
     _aaz_info = {
-        "version": "2023-07-01-preview",
+        "version": "2025-05-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.network/dnsresolverpolicies/{}/dnssecurityrules/{}", "2023-07-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.network/dnsresolverpolicies/{}/dnssecurityrules/{}", "2025-05-01"],
         ]
     }
 
@@ -129,12 +129,6 @@ class Update(AAZCommand):
             nullable=True,
             enum={"Alert": "Alert", "Allow": "Allow", "Block": "Block"},
         )
-        action.block_response_code = AAZStrArg(
-            options=["block-response-code"],
-            help="The response code for block actions.",
-            nullable=True,
-            enum={"SERVFAIL": "SERVFAIL"},
-        )
 
         domain_lists = cls._args_schema.domain_lists
         domain_lists.Element = AAZObjectArg(
@@ -230,7 +224,7 @@ class Update(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-07-01-preview",
+                    "api-version", "2025-05-01",
                     required=True,
                 ),
             }
@@ -333,7 +327,7 @@ class Update(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-07-01-preview",
+                    "api-version", "2025-05-01",
                     required=True,
                 ),
             }
@@ -410,7 +404,6 @@ class Update(AAZCommand):
             action = _builder.get(".properties.action")
             if action is not None:
                 action.set_prop("actionType", AAZStrType, ".action_type")
-                action.set_prop("blockResponseCode", AAZStrType, ".block_response_code")
 
             dns_resolver_domain_lists = _builder.get(".properties.dnsResolverDomainLists")
             if dns_resolver_domain_lists is not None:
@@ -502,9 +495,6 @@ class _UpdateHelper:
         action = _schema_dns_security_rule_read.properties.action
         action.action_type = AAZStrType(
             serialized_name="actionType",
-        )
-        action.block_response_code = AAZStrType(
-            serialized_name="blockResponseCode",
         )
 
         dns_resolver_domain_lists = _schema_dns_security_rule_read.properties.dns_resolver_domain_lists