+++ src/dns-resolver/HISTORY.rst
@@ -2,6 +2,12 @@
 
 Release History
 ===============
+1.1.0
+++++++
+* Bump from 2023-07-01-preview to 2025-05-01
+* Adds bulk domain list API for large domain list usage
+* Removes BlockResponseCode from DNS Security Rule input
+
 1.0.0
 ++++++
 * Bump from 2022-07-01 to 2023-07-01-preview